+++ tests/fixtures/api/v1/batches/00000000-0000-0000-0000-000000000001/cancel/jobs/_.PUT.json
